fix: 검색 결과 페이지 UX 개선 작업 #127
